+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- changed: Extended `onSubscribeAddresses` callback to optionally include checkpoints.
+
 ## 2.27.4 (2025-05-13)
 
 - fixed: Disable duress pin mistakenly disables pin-login for entire account.